+++ R/zzz.R
@@ -4,3 +4,8 @@
 #' @importFrom R6 R6Class
 #' @importFrom digest digest
 "_PACKAGE"
+
+.onLoad = function(libname, pkgname) {
+  # nocov start
+  backports::import(pkgname)
+} # nocov end