+++ tests/testthat/test_as_sqlite_backend.R
@@ -5,7 +5,6 @@ test_that("data", {
   expect_is(b, "DataBackendDplyr")
   expect_backend(b)
   expect_iris_backend(b)
-  b$disconnect()
 })
 
 test_that("DataBackend", {
@@ -13,12 +12,10 @@ test_that("DataBackend", {
   expect_is(b, "DataBackendDplyr")
   expect_iris_backend(b)
   expect_backend(b)
-  b$disconnect()
 })
 
 test_that("Task", {
   b = as_sqlite_backend(mlr3::mlr_tasks$get("iris"))$backend
   expect_is(b, "DataBackendDplyr")
   expect_backend(b)
-  b$disconnect()
 })